a few ARM and GCC13 build fixes

+++ 2.0/libdeflate/lib/arm/arm_cpu_features.c
@@ -80,7 +80,7 @@ static void scan_auxv(unsigned long *hwcap, unsigned long *hwcap2)
 				goto out;
 			}
 			filled += ret;
-		} while (filled < 2 * sizeof(long));
+		} while (filled < (int)(2 * sizeof(long)));
 
 		i = 0;
 		do {
@@ -93,7 +93,7 @@ static void scan_auxv(unsigned long *hwcap, unsigned long *hwcap2)
 				*hwcap2 = value;
 			i += 2;
 			filled -= 2 * sizeof(long);
-		} while (filled >= 2 * sizeof(long));
+		} while (filled >= (int)(2 * sizeof(long)));
 
 		memmove(auxbuf, &auxbuf[i], filled);
 	}
